+++ gamemodes/terrortown/gamemode/server/sv_admin.lua
@@ -48,10 +48,7 @@ end
 -- @param Player ply
 -- @realm server
 function PrintTraitors(ply)
-    ---
-    -- @realm server
-    -- stylua: ignore
-    if not IsValid(ply) or hook.Run("TTT2AdminCheck", ply) then
+    if not IsValid(ply) or admin.IsAdmin(ply) then
         ServerLog(Format("%s used ttt_print_traitors\n", IsValid(ply) and ply:Nick() or "console"))
 
         local pr = GetPrintFn(ply)
@@ -90,11 +87,10 @@ concommand.Add("ttt_print_usergroups", PrintGroups)
 function PrintReport(ply)
     local pr = GetPrintFn(ply)
 
-    ---
-    -- @realm server
-    -- stylua: ignore
-    if not IsValid(ply) or hook.Run("TTT2AdminCheck", ply) then
-        ServerLog(Format("%s used ttt_print_adminreport\n", IsValid(ply) and ply:Nick() or "console"))
+    if not IsValid(ply) or admin.IsAdmin(ply) then
+        ServerLog(
+            Format("%s used ttt_print_adminreport\n", IsValid(ply) and ply:Nick() or "console")
+        )
 
         for _, e in pairs(SCORE.Events) do
             if e.id == EVENT_KILL then
@@ -119,14 +115,10 @@ concommand.Add("ttt_print_adminreport", PrintReport)
 local function PrintKarma(ply)
     local pr = GetPrintFn(ply)
 
-    ---
-    -- @realm server
-    -- stylua: ignore
-    if not IsValid(ply) or hook.Run("TTT2AdminCheck", ply) then
+    if not IsValid(ply) or admin.IsAdmin(ply) then
         ServerLog(Format("%s used ttt_print_karma\n", IsValid(ply) and ply:Nick() or "console"))
 
         KARMA.PrintAll(pr)
-
     else
         if IsValid(ply) then
             pr("You do not appear to be RCON or a superadmin!")
@@ -197,10 +189,7 @@ local dmglog_save = CreateConVar("ttt_damagelog_save", "0", {FCVAR_NOTIFY, FCVAR
 local function PrintDamageLog(ply)
     local pr = GetPrintFn(ply)
 
-    ---
-    -- @realm server
-    -- stylua: ignore
-    if not IsValid(ply) or hook.Run("TTT2AdminCheck", ply) or gameloop.GetRoundState() ~= ROUND_ACTIVE then
+    if not IsValid(ply) or admin.IsAdmin(ply) or gameloop.GetRoundState() ~= ROUND_ACTIVE then
         ServerLog(Format("%s used ttt_print_damagelog\n", IsValid(ply) and ply:Nick() or "console"))
         pr("*** Damage log:\n")