+++ lua/terrortown/lang/fr.lua
@@ -915,7 +915,7 @@ L.pickup_pending = "Vous avez déjà pris une arme, attendez de la recevoir"
 L.tbut_help_admin = "Éditer les paramètres du traître"
 L.tbut_role_toggle = " Appuyez sur [{walkkey} + {usekey}] pour basculer ce bouton pour {role}"
 L.tbut_role_config = "Rôle: {current}"
-L.tbut_team_toggle = "Appuyez sur [SHIFT + {walkkey} + {usekey}] pour basculer ce bouton pour la team {team}"
+L.tbut_team_toggle = "Appuyez sur [SHIFT + {walkkey} + {usekey}] pour basculer ce bouton pour la {team}"
 L.tbut_team_config = "Team: {current}"
 L.tbut_current_config = "Configuration actuelle:"
 L.tbut_intended_config = "Configuration prévue par le créateur de la carte:"
@@ -2301,3 +2301,7 @@ L.decoy_help_primary = "Planter le leurre"
 --Restarting a round only restarts the current round so you can start over. Resetting the level clears everything so that the game starts new as if it is fresh after a map change.]]
 
 --L.label_button_level_reset = "reset level"
+
+--L.loadingscreen_round_restart_title = "Starting new round"
+--L.loadingscreen_round_restart_subtitle = "you're playing on {map}"
+--L.loadingscreen_round_restart_subtitle_limits = "you're playing on {map} for another {rounds} round(s) or {time}"