+++ lua/terrortown/lang/uk.lua
@@ -915,7 +915,7 @@ L.idle_popup_title = "Бездіяльність"
 --L.tbut_help_admin = "Edit traitor button settings"
 --L.tbut_role_toggle = "[{walkkey} + {usekey}] to toggle this button for {role}"
 --L.tbut_role_config = "Role: {current}"
---L.tbut_team_toggle = "[SHIFT + {walkkey} + {usekey}] to toggle this button for team {team}"
+--L.tbut_team_toggle = "[SHIFT + {walkkey} + {usekey}] to toggle this button for {team}"
 --L.tbut_team_config = "Team: {current}"
 --L.tbut_current_config = "Current config:"
 --L.tbut_intended_config = "Intended config by map creator:"
@@ -2301,3 +2301,7 @@ L.body_confirm_one = "{finder} підтверджує смерть {victim}."
 --Restarting a round only restarts the current round so you can start over. Resetting the level clears everything so that the game starts new as if it is fresh after a map change.]]
 
 --L.label_button_level_reset = "reset level"
+
+--L.loadingscreen_round_restart_title = "Starting new round"
+--L.loadingscreen_round_restart_subtitle = "you're playing on {map}"
+--L.loadingscreen_round_restart_subtitle_limits = "you're playing on {map} for another {rounds} round(s) or {time}"