+++ CHANGELOG.md
@@ -40,6 +40,10 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
   - Added a submenu to issue basic commands
 - Added a new `gameloop` module that contains all functions related to the round structure (by @Tim Goll)
 - Added a loadingscreen that hides the visible and audible lag introduced by the map cleanup on round change (by @TimGoll)
+- Added a voicebattery module that handles the voice battery (by @TimGoll)
+- Added `admin.IsAdmin(ply)` as a wrapper that automatically calls `GM:TTT2AdminCheck` (by @TimGoll)
+  - Made sure this new function is used in our whole codebase for all admin checks
+- Added `ENTITY:IsPlayerRagdoll ` to check if a corpse is a real player ragdoll (by @TimGoll)
 
 ### Changed
 
@@ -61,6 +65,8 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 - Cleanup and performance optimizations for marks library (by @WardenPotato)
 - Updated the Turkish localization file (by @NovaDiablox)
 - The level time now starts with the first preparing phase, meaning that idle on connect doesn't decrease the map time (by @TimGoll)
+- Minor cleanup and optimizations in weapon code (by @TW1STaL1CKY)
+- Now always properly checks if an entity is a true ragdoll to make sure no other props get ragdoll handling (by @TimGoll)
 
 ### Fixed
 
@@ -73,12 +79,15 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 - Fixed `DynamicCamera` error when a weapon's `CalcView` doesn't return complete values (by @TW1STaL1CKY)
 - Fixed Roundendscreen showing karma changes even if karma is disabled
 - Fixed the player's FOV staying zoomed in if their weapon is removed while scoped in (by @TW1STaL1CKY)
+- Fixed the player's FOV staying zoomed in with the binoculars if they're removed from you (by @TW1STaL1CKY)
 - Fixed weapon unscoping (or generally any time FOV is set back to default) being delayed due to the player's lag (by @TW1STaL1CKY)
 - Fixed overhead icons sometimes being stuck at random places (by @TimGoll)
 - Fixed a null entity error in the ShootBullet function in weapon_tttbase (by @mexikoedi)
 - Fixed a nil compare error in the DrawHUD function in weapon_tttbasegrenade (by @mexikoedi)
 - Fixed players sometimes not receiving their role if they joined late to the game (by @TimGoll)
 - Fixed weapon dryfire sound interrupting the weapon's gunshot sound (by @TW1STaL1CKY)
+- Fixed incendiaries sometimes exploding without fire (by @TimGoll)
+- Fixed scoreboard not showing any body search info on players that changed to forced spec during a round (by @TimGoll)
 
 ### Removed